Add defs and TDFs from Smithbox
This doesn't make any attempt to merge with or overrwrite existing
files. Where conflicts exist, it's not clear that either source is
more up-to-date or accurate than the other. In particular, Smithbox
often uses slightly different names for the same thing, and I don't
want to cause any tool breakages because of these differences.